python-service/requirements.txt
# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Image processing libraries
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Computer vision and ML libraries
scikit-image>=0.22.0

# PDF generation
reportlab>=4.0.0

# HTTP client for testing
httpx>=0.25.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Logging and monitoring
structlog>=23.2.0
psutil>=5.9.0

# File handling
python-multipart>=0.0.6

# Environment management
python-dotenv>=1.0.0